Refactor: JWT service extracted from authorization middleware
